chore: split the creator and preview into multiple webpack configs
