+++ sites/fast-creator/app/app.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter, Route } from "react-router-dom";
 import Creator from "./creator";
-import Preview from "./preview";
 
 /**
  * Create the root node
@@ -12,7 +11,6 @@ export default class App extends React.Component<{}, {}> {
         return (
             <BrowserRouter>
                 <Route component={Creator} exact={true} path="/" />
-                <Route component={Preview} exact={true} path="/preview" />
             </BrowserRouter>
         );
     }
