+++ sites/fast-creator/app/preview/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr" prefix="og: https://ogp.me/ns#">
+    <head>
+        <meta charset="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title><%= htmlWebpackPlugin.options.title %></title>
+        <%= htmlWebpackPlugin.options.manifest %>
+        <meta
+            name="title"
+            property="og:title"
+            content="<%= htmlWebpackPlugin.options.title %>"
+        />
+    </head>
+    <body>
+        <div id="root"></div>
+    </body>
+</html>
