+++ sites/fast-creator/app/preview/webpack.common.js
@@ -1,0 +1,73 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+/* eslint-disable no-undef */
+const path = require("path");
+const { CleanWebpackPlugin } = require("clean-webpack-plugin");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+
+const appDir = path.resolve(__dirname, "./");
+const outDir = path.resolve(__dirname, "./www");
+
+module.exports = {
+    name: "preview",
+    entry: {
+        main: path.resolve(appDir, "index.tsx"),
+    },
+    resolve: {
+        extensions: [".ts", ".tsx", ".js"],
+    },
+    output: {
+        path: outDir,
+        publicPath: "/",
+    },
+    module: {
+        // where we defined file patterns and their loaders
+        rules: [
+            {
+                test: /.tsx?$/,
+                use: [
+                    {
+                        loader: "ts-loader",
+                    },
+                ],
+            },
+            {
+                test: /\.(svg|png|jpe?g|gif|ttf)$/i,
+                use: {
+                    loader: "file-loader",
+                    options: {
+                        esModule: false,
+                    },
+                },
+            },
+            {
+                test: /\.css$/,
+                use: [
+                    {
+                        loader: MiniCssExtractPlugin.loader,
+                        options: {
+                            hmr: process.env.NODE_ENV === "development",
+                        },
+                    },
+                    {
+                        loader: "css-loader",
+                    },
+                ],
+            },
+            {
+                test: /message\-system\.min\.js/,
+                use: {
+                    loader: "worker-loader",
+                },
+            },
+        ],
+    },
+    plugins: [
+        new MiniCssExtractPlugin(),
+        new HtmlWebpackPlugin({
+            title: "FAST Creator Preview",
+            inject: "body",
+            template: path.resolve(appDir, "index.html"),
+        }),
+    ],
+};
