+++ sites/fast-creator/app/utilities/index.ts
@@ -1,1 +1,2 @@
 export * from "./resolver";
+export * from "./shared";
