+++ sites/fast-creator/app/utilities/shared.ts
@@ -1,0 +1,6 @@
+/**
+ * These are shared configs or items between the Preview application and the Creator application
+ * which are compiled separately.
+ */
+
+export const designTokensLinkedDataId: string = "design-tokens";
