+++ sites/fast-creator/webpack.dev.js
@@ -1,16 +1,17 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 /* eslint-disable no-undef */
 const merge = require("webpack-merge");
-const baseConfig = require("./webpack.common");
+const baseConfigs = require("./webpack.common");
 
-module.exports = merge(baseConfig, {
-    devServer: {
-        contentBase: "./src/public",
-        open: true,
-        port: 7777,
-    },
-    mode: "development",
-    output: {
-        filename: "[name].js",
-    },
+module.exports = baseConfigs.map(baseConfig => {
+    return merge(baseConfig, {
+        devServer: {
+            open: true,
+            port: 7777,
+        },
+        mode: "development",
+        output: {
+            filename: "[name].js",
+        },
+    });
 });
