+++ sites/fast-creator/webpack.prod.js
@@ -1,33 +1,34 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 /* eslint-disable no-undef */
 const merge = require("webpack-merge");
-const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const OptimizeCSSAssetsPlugin = require("optimize-css-assets-webpack-plugin");
 const TerserJSPlugin = require("terser-webpack-plugin");
-const baseConfig = require("./webpack.common");
+const baseConfigs = require("./webpack.common");
 
-module.exports = merge(baseConfig, {
-    mode: "production",
-    optimization: {
-        minimizer: [new TerserJSPlugin({}), new OptimizeCSSAssetsPlugin({})],
-        splitChunks: {
-            chunks: "all",
-            maxInitialRequests: 100,
-            cacheGroups: {
-                vendor: {
-                    test: /[\\/]node_modules[\\/]/,
-                    name: module => {
-                        const packageName = module.context.match(
-                            /[\\/]node_modules[\\/](.*?)([\\/]|$)/
-                        )[1];
-                        // npm package names are URL-safe, but some servers don't like @ symbols
-                        return `npm.${packageName.replace("@", "")}`;
+module.exports = baseConfigs.map(baseConfig => {
+    return merge(baseConfig, {
+        mode: "production",
+        optimization: {
+            minimizer: [new TerserJSPlugin({}), new OptimizeCSSAssetsPlugin({})],
+            splitChunks: {
+                chunks: "all",
+                maxInitialRequests: 100,
+                cacheGroups: {
+                    vendor: {
+                        test: /[\\/]node_modules[\\/]/,
+                        name: module => {
+                            const packageName = module.context.match(
+                                /[\\/]node_modules[\\/](.*?)([\\/]|$)/
+                            )[1];
+                            // npm package names are URL-safe, but some servers don't like @ symbols
+                            return `npm.${packageName.replace("@", "")}`;
+                        },
                     },
                 },
             },
         },
-    },
-    output: {
-        filename: "bundle/[name].[contenthash].js",
-    },
+        output: {
+            filename: "bundle/[name].[contenthash].js",
+        },
+    });
 });
